Replaces fastlane render w/ redirect to download
The fast lane page used to just render the download template,
which didn't show the newly created unique URI (share url)
in the address bar.

This commit modifies the fast lane view to redirect to the download
view using an existing or newly created build id.

In the download view we now check if the build was a fast lane build
and if it was built in the same session to know if we also have
to render the download keys section of the template.